Fix new line at the end of TypeScript code generation (#50)
It turned out that TypeScript generated files does not end with a new
line character. The issue lies in the fact that
`generator.printOnNewline()` method was used that accepts a string
parameter. If the string is `falsy` it does nothing. The fix is to use
`printNewline()` method instead.

The tests were passing because the tag helper `stripIndent` is deleting
the new line charter when the line is ‘empty’. The tests are fixed by
appending the new line character to the expected value.