Redo filtering out commented out gql strings
We want to filter out gql strings that are completely commented out,
example:

```
// gql`
// query UserProfileView {
//  me {
//    id
//    uuid
//    # TODO https://app.asana.com/0/52068106376728/221203365474362
//    role
//  }
//}
`
```
(and the block comment version)

Previously, we had a regex: '\s*`([^`/]*)`, where we included `/` in the
`not` part of the regex. Unfortunately, filtering out comments with
regex is difficult and I don't think it's possible to do it 100% of the
time.

This regex however filters otu gql strings like
```
gql`
query UserProfileView {
  me {
    id
    uuid
    # TODO https://app.asana.com/0/52068106376728/221203365474362
    role
  }
}
`
```

Where the comment in the gql string is valid.

This changes it so it passes the source file through babel.transform
with `comments:false` to remove comments before extracting gql strings.