Emit JSDoc comments for TypeScript types and enums
Many IDEs that work with TypeScript will use JSDoc comments on types,
fields, and enum variants in hover and completion contexts. Changing
the generated code to wrap the `description` property in a JSDoc comment
instead of a regular comment or postfix comment makes those IDEs surface
the GraphQL description to the developer as they're working.

Fixes #384