Use output arg instead of --localDirectory
Placing Flow and TypeScript files in directories relative to their source files is now the default behavior. By default, the directory name used is `__generated__`, which can be overridden by setting the `output` (first) cli arg.

To use the `output` to specify a single locaation the new `--outputRelativeToCWD` flag can be set.

To continue placing generated files immediately adjacent their source files (which is the current behavior, albiet only for a couple days and arguably undesireable) `output` can be set to an empty string. While available, I suspect this feature will never be used, due to the problems outlined in #468 and #469.