Restore ability to put generated files in a directory relative to source (#470)
* Fix broken precommit script.

It appears that the precommit script was incorrectly defined in the packages/apollo-cli/package.json file, which was making it impossible to commit without a --no-verify flag.

I fixed the configuration, and applied the format changes to the to-be-changed files in this commit.

* Add `localDirectory=name` flag to TypeScript/Flow generators.

Before the change from `apollo-codegen` to `apollo`, flow type definitions were output into a `__generated__` directory, next to their respective source files. The new `localDirectory` flag restores support for this kind of workflow by accepting a custom directory name:

```sh
apollo codegen:generate --target=flow --localDirectory=__generated__
```

This flag is ignored if `--output` is specified.

* Use output arg instead of --localDirectory

Placing Flow and TypeScript files in directories relative to their source files is now the default behavior. By default, the directory name used is `__generated__`, which can be overridden by setting the `output` (first) cli arg.

To use the `output` to specify a single locaation the new `--outputRelativeToCWD` flag can be set.

To continue placing generated files immediately adjacent their source files (which is the current behavior, albiet only for a couple days and arguably undesireable) `output` can be set to an empty string. While available, I suspect this feature will never be used, due to the problems outlined in #468 and #469.

* rename outputRelativeToCWD to outputFlat