Parse files with recast and find templates in AST
Fallback on the naive regext search for gql` in the cases that the
parsing fails. This also removes the code that chased down imports and
attempted to resolve them with variable/import analysis. This strategy
proves to be too fragile when attempting to construct an interpolated
string. There are too many different ways to import values. Additionally
in the future world of import(), which returns a promise, would be
impossible to support.

* import apollo-engine-reporting for AST manipulation

* remove unused variables

* continue traversing AST for nested template literal