Switch manifest to use SHA-256 hexadecimal rather than SHA-512 base64.
While I'm at breaking changes, we don't have a compelling enough reason to
support SHA-512 just yet.  The ask for 512 was previously my own since it
could have a longer lifespan (and is sometimes faster) and the base64
encoding was to reduce the rather long length of SHA-512's hex encoding
(which is 128-bytes rather than the 64-bytes which result from SHA-256's
hex encoding).