Rebuild the CLI
First off, if you have reached this point in the git history, I'm sorry.
Hopefully this will never be found because I will have had time to split
this commit into about 10 smaller commits, but alas Summit is days away
and all of this needed to work. Forgive me team as I have sinned and
created a monster commit. It was to ease the nasty rebase awaiting
me from Maritjn and because there was almost no parital state at
any working time.

- with deep sadness, jbaxleyiii