Add event-stream as a dep and lock it (security issue)
As identified in https://github.com/dominictarr/event-stream/issues/116,
`event-stream` has a major security issue (malware injection) in
version 3.3.6 (thanks to `flatmap-stream` version 0.1.1).
`event-stream` 3.3.6 is referenced as a child dep in this
project, through `tsc-watch` and `vscode-apollo` / `vscode`.

This commit adds `event-stream` as a top level dependency, and
locks it to the most recent version that excludes `flatmap-stream`
(version 3.3.4).

This should work for now, but ultimately `tsc-watch` and
`vscode` should be updated to newer versions, that address this
issue (since their child deps are the problem). Both projects
have yet to submit fixes to this problem.