Regenerate `package-lock.json` to fix a number of transitive deps.
In light of the `event-stream` vulnerability, which was included in the
`apollo-vscode` extension through Microsoft's own dependency through the
`vscode` npm (https://npm.im/vscode), this is an overly cautious update to
the `package-lock.json` to avoid a number of packages who have had their
sources removed from the registry.

I've carefully reviewed the diff of this `package-lock.json` to try any
ensure no other breaking changes are introduced, but given the severity of
the security issue, it's best to move forward on this now and act on any
residuals later.  The most notable changes here to pay attention to are the
`vscode*` extensions (related to `*gulp*` packages) and the very explicit
pinning of `event-stream` within those packages to semver `"3.3.4"` (i.e.
_exactly_ 3.3.4, which is the last known non-vulnerable version.