Add fallback for git branch calculation
In some ci's, the git branch will not be present in the environment with
the default that env-ci uses, so we need to provide a fallback. The
fallback is either the PR's branch if present or the .git branch. This
branch is used inside of the Engine frontend to display more information
about the schema changes.

See https://github.com/pvdlg/env-ci#caveats for a longer list of cases
when the branch will not be present in the ci environment