Filter out 'undefined' definitions before checking the length.
It's possible to leave a whole query behind when filtering by directive. In that case, we want to exclude it altogether.
In the current state, definitions can === [undefined], giving a length of 1 and passing this check. By filtering for
truthy values only, we won't include empty operations in the result of this function.