Add `esModuleInterop: true` to `apollo-graphql`'s `tsconfig.test.json`.
Tests for `apollo-graphql` are working again as of this commit, but the
tests highlighted interop failures with regards to `esModuleInterop` since
the testing configuration for this repository seems to be partially
configured to use different configurations.  We should probably fix that!

Since this setting isn't inherited from a monorepo root configuration, this
needs to be defined in both the package's `tsconfig.json` and its
`tsconfig.test.json`.