Fix tests for `apollo-graphql` by adding `jest` config.
I observed that the tests weren't running for the `apollo-graphql` package
when reviewing the work in #1018.

Upon investigating further (via `jest --debug`), it seems that the Jest
configuration was missing the appropriate configuration for it to pickup the
extensions of `ts` and `tsx` within the `apollo-graphql` repository.

It was only then that I realized that this repository still has `jest`
configuration defined in each of its `package.json` files, rather than a
monorepo-root `jest.config.base.json` file.  That's fine, but it caught me
off guard.

I do think that we should try to DRY-up this repetition and align this
across teams and repositories as soon as possible because it's difficult to
start building per-package exceptions which will quickly become hard to
manage.  For now, this will get tests working for `apollo-graphql` within
the scope of the PR that brings it into the `apollo-tooling` repository,
which seems necessary to tick the box for shipping that PR.