Update `child_process.spawn` usage to pass empty list in `args` parameter.
While Node.js uses the heuristic of "If you're passing an array as the
second argument to `spawn` they must be arguments.  If the second argument
is instead an object, they must be the options.", the Node.js v8 types don't
seem to be aware of that yet.

I think it's worth using the Node.js 8 types on this repository until we
drop support for Node.js 8.  They are still maintained, though some of the
fixes don't get backported.

Another option for this, since this seems to be a direct wrapper over
`node`, would be to just spawn `node` from the npm command, but I'm not sure
why the decision was made to go this direction originally.

cc @trevor-scheer

Node.js 10.x Ref:
https://nodejs.org/docs/latest-v10.x/api/child_process.html#child_process_child_process_spawn_command_args_options