Fix __typename addition for InlineFragments
We need operation registration's __typename adding to match how
Apollo Client does this in order for query hashes to match.

In order to achieve parity here, I've added the __typename field
to InlineFragments as well. While the __typename on an interface
may be redundant (see new tests for example), it's also a safer
change for me to recommend than anything within Apollo Client.