Pin graphql to the ~14.2.x range
An incoming graphql release (14.3.1) is going to break at least
`apollo client:check` as it did in the past (confirmed this
manually with graphql's master branch). This dependency restriction
will prevent this for now, without us needing to find the root of
the issue. This is preventative for now until we have the resources
to investigate and resolve the actual cause of the problem.

For more details and full context, see:
https://github.com/graphql/graphql-js/commit/183ff32bee4bc23eb23657f79f414c2400ecb06a#r32971387