client:push add check for parse failures
Currently the client project returns operation that parse correctly.
This strategy swallows parse failures. This updates client:push to
surface the unparseable operations.

Ideally this would be a more comprehensive solution that solves the
issues for extract as well as check, so some feedback on a better
solution is highly encouraged