Recognize .graphqls as graphql file
The team I work in intends to use the `.graphqls` file extension to declare queries, mutations and fragments, because some of its developers claimed it has wider support in some extensions for IDEs like Webstorm. 

I noticed that apollo-codegen does not select `.graphqls` files to generate types, but the extension does get mentioned at several places in the apollo-tooling repo, including apollo-codegen-core and apollo-language-server - right beside the extensions `.graphql` and `.gql`. 

Including this change (simulated with node + devtools) will make apollo codegen select `.graphqls` files. Is there a particular reason not to include this extension here? When I googled the use of the extensions, I did not get a clear or official answer on when to use which. 

If not, may this change be merged?