Change usages of tag to also look to variant option
I'm not entirely sure what's going on here with respect to tags,
variants, etc. and how the config interacts with the flags, and it's not
like this is tested, so this is somewhat of a shot in the dark. Feel
free to revert these commits, but I think a follow-up should add
'variant' as an option in the apollo.config.js and throw an error when
both variant and tag are set. In general, I think we should be
consistent in looking to `flags.variant || flags.tag || config.tag ||
'current', which I think `config.tag` started to do by defining a getter
and having it always fall back on 'current', but there was some black
magic there, so I kind of just threw certain truth at the problem rather
than thinking through what a nice way to share this logic would be.
Hopefully this is at least minorly useful, even if it needs to be undone
and refactored.