Add overrideComposedSchema option to service:push
By default, it is forbidden to push a schema directly to a federated
graph. If users try to push a schema directly, the API will return
errors unless 'overrideComposedSchema' is specified in the arguments.
This change adds the boolean flag that corresponds to this "force push"
to the CLI.

It also runs lint-fix, unfortunately in the same commit.