fix(language-server): remove fragment definitions when only referenced in @client field
If we spread a fragment inside of a `@client` field, we remove that
fragment spread. If we don't use that fragment anywhere else then we should
remove the fragment definition.

It _is_ possible that a fragment is used on a @client field and non-client
field if they are fragments on an interface and not and object.