Add `includeMergedTags` to the `lerna.json` to support release branches.
Following the merged tags is important to Lerna since it uses Git tags and
branches to follow the release flow.  This is arguably not necessarily in
this repository since it doesn't heavily leverage release branches, but it
won't hurt to add it and it provides consistency with other repositories
that we have (Apollo Server, at the very least).