swift: fix initializers with properties named `self`
When a type has a field `self`, the synthesized initializers need to
avoid using `self` as the internal parameter name, instead permuting it.
By default we pick `_self`, but we also ensure this doesn't conflict
with any other declared properties; in the event there is a conflict, we
append `_` until it's unique.

This makes the test introduced by the previous commit start passing.

Fixes #1530.