Install `graphql-js` into `devDependencies` at the monorepo level.
Currently, even though most of this repository currently specifies `graphql`
as a peer-dependency, the only reason it's installed at the monorepo root,
and thus available for use by all packages, is because it's a direct
dependency of `apollo` and `apollo-language-server`.

This is causing it not to be upgraded to the most recent version by
Renovate.  In my own case, I needed the latest version for TypeScript
typings for `PossibleTypeExtensions` which were not ever officially
published in the DefinitelyTyped typings (likely due to the method being
experimental) but are now part of the `graphql` package itself thanks to the
work that @JacksonKearl did.