Improve performance of client directive validation (#1559)
* This commit seeks to improve the performance when determining
if a field is missing any @client directives when using local state
with Apollo Client. Instead of recursing over the entire document for
each field, we walk the tree once while checking the schema for any
possible client side annotations. This PR is an initial take at
improving this workload but doesn't complete the work.

Still to be done includes supporting @client validation on fragment
spreads

* fix TS error on failure to clean up removed function

* Improve startup performance of CLI by removing extra validation

Historically, the language server was designed to be used just with the
VS Code extension (and other editors) where, after loading the project,
the first thing we want to do is validate all of the documents and
schema. However, we resuse the base projects (client / server) in the
CLI to share common work especially around document management. Due to
this, when the CLI starts up, its runs a potentially expensive
validation step regardless of if the command has anything to do with
validation. Often this is a no-op because of a missing schema, but given
the correct env variables that loading can still happen automatically
(something to clean up later). This change moves the logic of "auto
validate" to the concept of a workspace which is used in long running
processes (currently the VS Code but soon the CLI watch mode) and keeps
the base project class clean. It also removes the hack in place that
helped to highlight this in the recent added test suite for client
validation.

* update changelog

* Pragmatic take on improving codegen efficiency

This simplifies the watch and non watch modes by only calling write
within the watcher and not relying on the underyling documentChange
events from the project base. It also add a *VERY* simple exclude
support which is far from the right solution but prevents more loops
then currently happen

* Add gitContext to checkPartialSchema mutation (#1560)

* add gitContext to federated checkPartialSchema mutation

* Clarify changelog and fix typo