Use service endpoint in service:check if provided
This does a number of things, but primarily it uses the
`service.endpoint` configuration to provide an endpointSchemaProvider
when running `service:check` if available. This may require more careful
consideration for how to support a simultaneous `service` and `client`
key in the configuration, and we may want to have a
`serviceSchemaProvider` and `clientSchemaProvider` to support this use
case, but this changes `service:check` in particular to work within the
context of a client configuration. This is important because clients may
want to diff different running servers or environments. However, we will
need to guarantee that `client:codegen`, when uses, continues to point
to the client schema provider in lieu of the service schema provider.

Additionally, this changes `config.name` to the more appropriate
`config.graphId`