Only load schema on startup if requested
In general, we do not want to load the schema on startup when we are
testing, since it would require nocking out calls to the graph manager
API. Instead, we have the Command.ts _not_ load the schema on startup,
and only have the VS Code extension do so. This requires calling
loadSchema() when you want to access it.