Don't treat GraphQL's ID type as builtin for TypeScript
If you want to attach a custom type in TypeScript for the IDs of GraphQL
objects, you cannot do it even by passing `--passthroughCustomScalars`.
While technically, ID is not a custom scalar, we may find a custom type
to represent it with in TypeScript.

This change introduces a backwards incompatibility, by changing the
default custom scalar types (when no `--passthroughCustomScalars` is
given) from `any` to `string`. I believe this is a fair change as the
swift codegen adapter already behaves like this and by spec[1], the
scalars in GraphQL can always be represented at least by a string.

[1]: https://graphql.github.io/graphql-spec/June2018/#sec-Scalars