Revert "Unify and simplify jest config (#1612)"
This reverts commit b0332ad9675285e25f2e6c38fa3f12a398b9a973.

After trying to add some tests to the apollo-graphql project,
I realized that the test suite was being ignored by Jest completely.
Along with a few other suites for some reason that I can't seem
to figure out with the amount of time I had available to me to
investigate.

The skipping of these tests is a result of PR #1612 mentioned above.
This PR reverts cleanly and reintroduces the tests that were being
skipped. It would be nice to go back and reattempt to land this at
some point, just be cautious of the test count mysteriously dropping.