Change non-existent service message to be correct
Currently, the message we provide to users for querying a service that
does not exist (or that they do not have access to) looks up the service
ID from the key rather than using the actual variable that it passed to
AGM when looking it up. This changes such messages to always use the
actual service ID that they were searching for.

Specifically, this is to support user tokens, where the graph in
question may not, in fact, be in the key, but instead by provided by
apollo.config.js