Update "no service found to link..." error (#1847)
* Update "no service found to link..." error

This updates the error message that occurs when a config.service.name is
required in order to execute a command. Before this commit, there were
two distinct messages that were used in commands to communicate this
error to the user, which were slightly different in their wording. This
updates the wording to provide an explanation as well as next steps on
how to remedy the issue. Additionally, it uses the `CLIError` class in
order to print out an error in a "friendly" way without a stack trace,
as recommended by the @oclif/errors package we are using: https://oclif.io/docs/commands#thiserrormessage-string--error-options-code-string-exit-number

* Remove unused imports

* Update packages/apollo/src/utils/errors.ts

Co-Authored-By: Jake Dawkins <dawkinsjh@gmail.com>

Co-authored-by: Adam Zionts <adam@meteor.com>
Co-authored-by: Jake Dawkins <dawkinsjh@gmail.com>