Create --variant flag, deprecate --tag flag (#1849)
* Make newlines (optional commit)

* Change references of Apollo Engine in flags

These default flags were still referencing Apollo Engine instead of
Graph Manager (even though some were hidden). Additionally, this updates
the docs to be more accurate. Happy to take this commit into another PR
if you'd like!

* Support variant flag in service:list command

This change supports the variant flag (with character -v) in addition to
the "tag" flag, which allows us to start using the new terminology. The
'--tag' flag is still supported, and there is verification that the two
sibling flags cannot be used together.

* Give deprecation warning to --tag users

* Rename errors util to sharedMessages

* Move tag flag deprecation warning to sharedMessages

* Support the variant flag in service:check

Also remove an unused variable in service:list

* Update service:delete to use variant flag

Also update description of the flags

* Add support to service:download to use --variant

Also simplify an import in apollo-language-server
Also add a XXX comment to service:download to illustrate a problem

* Change magical config tag to use variant

Additionally, this updates the text output in the service:check command
to consistently use graph@variant everywhere

* Rename magical config.tag to config.variant

This is a find and replace all, and I verified every command still works
after this (both with --tag and --variant). Will post screenshots of how
everything looks in the PR :)

* Update snapshots

A previous commit added `@variant` everywhere, and these snapshot
updates should be ONLY that.

* Put all handling into the magical config.variant

* Replace all flags.tag references to config.variant

* Add variant to the client: commands

All of the client commands share common flags, so no further changes are
required. Also, all of them use config.variant, so yay! :)

* Warn about deprecation at magical config.variant

Remove the deprecation warning from specific callsites and rely on the
deprecation warning appearing when the config.variant value is set

* Support for service:push variant

Also, this changes the default value of the flag, but it does NOT change
the behaviour, since `config.variant` always defaults to "current" if
flags and configuration are not set. The reason this was done is because
flags.tag will be non-null if there is a default, which causes the
deprecation message to always print

* Changes word "tag" in some user-facing messages

* Some small tweaks in wording

Mainly renaming service to graph and "schema" to service, which felt a
little meh, but more in line with the name of the command

* Update changelog

* Updates snapshots to match new messaging

This should only include changes to some minor message changes in
service:check

Co-authored-by: Adam Zionts <adam@meteor.com>