Support APOLLO_KEY, deprecate ENGINE_API_KEY
This supports a more modern name for the API key and adds deprecation
messages for the (now legacy) ENGINE_API_KEY. Much in the same spirit
as https://github.com/apollographql/apollo-tooling/pull/1849, the goal
is to modernize our documentation to no longer use the out-of-date name
"engine". This also modifies tests to support the new version and adds a
test that ensures an error is thrown when a user sets _both_ keys.