Fix silent codegen errors on syntax errors
During document parsing, syntax errors are silently
added to the document, but errors are never thrown.

This PR checks at the _command-level_ for these errors
and if they exist, prints them to the user.

I didn't want to change at the document setup level, because
I have a feeling the silent erroring was intention for the sake
of the operation registry or something. So I added in an error
where it was immediately causing an issue, in the client
codegen command.