Change json-modern target to use legacy IR
The legacy IR contains information that codegen in swift needs to
work properly. When building the json-modern target, I didn't
account for everything that was dropped from the original json
output :sad:. This commit changes the json-modern codegen target
so it uses the legacy IR generation with the exposeTypeNodes option
set to true (which passes it to the new ir codegen since legacy
uses that under the hood lol).