fix(apollo-language-server): Confirm file path is a file in `fileDidChange`
If there is a directory named like a file, meaning the path ends in
something like `.ts`, then the glob in an apollo config could pick up
directories. We need to make sure we are trying to open a file before we
run `readFileSync` or having this weirdly named path will cause the CLI to
crash.

FWIW We're seeing this in AGM right now because
[Cypress](https://www.cypress.io/) creates screenshots for tests with the
test filename as the directory name in the screenshots directory.