[AR-1691] The CLI does not need to know the frontend URL
Currently the CLI has configuration (via `engine.frontend` in a config file or
the --frontend flag to a few commands) to control what Engine frontend the user
is using. This is primarily set by Apollo employees using staging servers. This
is in addition to setting `engine.endpoint` to control which GraphQL server it
talks to.

This is used for two things:
- passing arguments to a couple ServiceMutation fields
- printing links in `services:list`

As of recently, this no longer is needed:
- those ServiceMutation fields now ignore the `frontend` argument and just use
  the correct frontend for the server processing them
- we expose a Query.frontendUrlRoot field that `services:list` can use

So this change completely removes recognition of this config field as well as
the flag. (Note that it is not an error to have unrecognized fields in your
config file, though it is to pass an unknown flag.) While this is technically
backwards-incompatible, we don't expect that this undocumented feature is used
much other than by Apollo employees.