Revert jest updates introduced by 9db1157
The update to jest@25 in the mentioned commit breaks local dev
testing for a currently unknown reason, but it seems related to
the filesystem. The current state of things mysteriously passes
in CI, but roughly 19 tests fail consistently in multiple dev
environments. For now, this commit resolves the issue.

Note that all jest dependencies are now pinned in the renovate
config to prevent these auto-updating again.

cc @abernix @JakeDawkins