Update markdown and json options to respect exit codes
Before this change, the --json and --markdown option would early return
and always exit with 0. This changes them to exit with non-zero codes
when the check is failing. The code that is accommodating these flags
could really afford to be abstracted and cleaned up quite a bit (in
fact, this whole file could), but it's a bit gnarly of a weekend
project, so this is just a fix for the bug.

Additionally, this updates checks to always exit with code 4 on
composition errors and code 1 on other breaking changes.