Temporarily disable the "Test VSCode" tests.
Currently, these are segmentation faulting both locally and in CI because of
transitive, non-deterministic dependency updates.  The first culprit I would
assume is the now deprecated `vscode` package which is responsible for
installing and invoking the tests with VSCode per instructions we used to
instrument them some time back.

That package is now deprecated and there are migration steps offered on
the package's npm page: https://npm.im/vscode

We should do that, but it's important that we get a release out right now
of the `apollo` CLI and that does not need to pass VSCode tests.