deps: Stop using fork of `graphql-language-service-*` packages.
Originally, we forked the `graphql-language-service-*` packages to avoid a
dependency on `graphql-config` and `graphql-import` which were included as
direct dependencies of those packages (and problematic because of their OWN
dependency trees, at the time).

That is no longer the case in the upstream packages.  Furthermore, the
TypeScript types that we wrote to utilize those packages (which were at the
time written in Flow) are no longer necessary since the project has been
ported to TypeScript.

This is an attempt to switch them out for our own packages.