Cease using ES features which don't exist natively on our target platforms
Honestly, this whole attempt to use modern ECMAScript features before they
were ready and truly supported in the runtimes we support was I think a
great experiment, but not worth the struggle.  We still support Node.js 10
for a couple more weeks, but this is a constraint for me today and it's
(being, generally the `apollo-server-env` and the `apollo-env` package's
polyfills) have been a constraint _many_ times over.

Put another way, I don't think the Pandora's box of "polyfills" is worth
it for things like `Array.protytype.flat`, which are not so inherently
complex that they can't be abstracted into a small function.

The `core-js-pure` package does a pretty good job offering this
functionality without necessitating us in-house/vendor it ourselves, but it
- sadly - does not provide TypeScript types, unlike the `core-js` core
itself.