ci: Disable single VS Code test because of failing VS Code test harness.
The "VS Code" testing that we have setup, while something that has been
implemented for a while, is relatively under-used; there is only a single
test today.

That testing framework was put in place some time ago, using a package and a
technique that is now-deprecated and should be upgraded to a new pattern:

https://code.visualstudio.com/api/working-with-extensions/testing-extension#migrating-from-vscode

Unfortunately, something about the existing `vscode` extension's technique
of automatically downloading VSCode versions from their infrastructure just
stopped working randomly with no changes at all to the dependencies of this
package.  I found this through a long git-bisect process which revealed that
even old commits years ago and even the commit that first introduced the
testing are now failing.

Since there is only one test which tests a very small unlikely to be touched
statusbar component of the VS Code extension, I am disabling the VS Code
testing in CircleCI since it's complicating all PRs right now and is not
really providing much in the way of confidence around the overall stability
of the VS Code extension anyhow.