chore(deps): Update to npm v7 and enable `engine-strict`
Essentially, the same updates as
https://github.com/apollographql/federation/pull/708

The intent of introducing `engine-strict` is to make sure that the users of
this monorepo are not led astray into using an unsupported version of
Node.js or npm that might cause unexpected side-effects. (Running the wrong
version of `npm` is the most motivating reason today!)

Users who wish to run this monorepo with a version of Node.js and npm that
are not specified in the `engines` range may still do so by running `npm
install --no-engine-strict` to acknowledge that there might be
unexpected/untested results/outcomes from their actions.