chore(deps): Update to npm v7, enable `engine-strict`, test Node.js 16 (#2302)
* chore(deps): Update to npm v7 and enable `engine-strict`

Essentially, the same updates as
https://github.com/apollographql/federation/pull/708

The intent of introducing `engine-strict` is to make sure that the users of
this monorepo are not led astray into using an unsupported version of
Node.js or npm that might cause unexpected side-effects. (Running the wrong
version of `npm` is the most motivating reason today!)

Users who wish to run this monorepo with a version of Node.js and npm that
are not specified in the `engines` range may still do so by running `npm
install --no-engine-strict` to acknowledge that there might be
unexpected/untested results/outcomes from their actions.

* Stop testing Node.js 10

Node.js 10 is officially end-of-life as of the end of April 2021.

While these packages may and should continue to work for those that are
stuck on old versions, those versions are no longer receiving security
updates and should be prioritized in terms of getting upgraded to newer
versions.

* chore: Update to `cimg` images, in preparation for Node.js 16 testing.

The newer images are smaller and suggested to be widely cached.  In
practice, the older images are widely cached because they're still more
prevalent but happy to use the newer images, particularly since the older
images aren't getting as much in the way of updates anymore (e.g., there are
not any Node.js 16 images for it.)

Similar in spirit to https://github.com/apollographql/federation/pull/713/commits/a46cdecf183dc74b014838004df3a73e82829b36

* chore: Start testing Node.js 16

* chore: install npm@7 and Node.js 14 on Windows in the easiest way I know how

Completely copied from the federation repository.  We might want to make an
Orb.